requirements.txt
fastapi[all]<=0.99.1
uvicorn[standard]
python-multipart
gunicorn
numpy
tritonclient[all]==2.45.0
sentry-sdk>=1.9.6
requests>=2.27.1
pydantic>=1.9.0
imageio>=2.18.0
blend-modes>=2.1.0
PyWavelets>=1.3.0
catboost>=1.0.6
redis
imutils>=0.5.4
pyclipper
shapely
ftfy
regex
Wand
svglib
dataclasses-json<=0.5.9
elastic-apm
tqdm
natsort
opencv-python-headless==4.5.3.56
scikit-image==0.18.3
matplotlib==3.4.3
scipy==1.7.1
timebudget==0.7.1
joblib==1.0.1
Pillow==8.3.2
scikit-learn==0.24.2
dill==0.3.4
diskcache==5.6.3
protobuf
trimesh
open3d
plyfile